--- stv_project_1.1.1/stv_project_1.1.1/stv_project_1.1.1.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "self:stv_project_1.1.1.xcodeproj">
-   </FileRef>
-</Workspace>
