--- stv_project_1.1.1/stv_project_1.1.1/stv_project_1.1.1/Assets.xcassets/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
